+++ tests/ViewTest.php
@@ -19,7 +19,6 @@ use Yiisoft\View\Tests\TestSupport\TestTrait;
 use Yiisoft\View\Theme;
 use Yiisoft\View\View;
 use Yiisoft\View\ViewContextInterface;
-
 use function crc32;
 use function dechex;
 use function file_put_contents;
@@ -184,20 +183,27 @@ PHP
         $view = $this
             ->createViewWithBasePath($this->tempDirectory)
             ->withContext($this->createContext($this->tempDirectory));
-        file_put_contents("$this->tempDirectory/file.php", 'Test');
-        file_put_contents("$this->tempDirectory/file.tpl", 'Test');
-        file_put_contents("$this->tempDirectory/file.txt.php", 'Test');
+        file_put_contents("$this->tempDirectory/file.php", 'Test php');
+        file_put_contents("$this->tempDirectory/file.tpl", 'Test tpl');
+        file_put_contents("$this->tempDirectory/file.phpt", 'Test phpt');
+        file_put_contents("$this->tempDirectory/file.txt.twig", 'Test txt');
 
-        $this->assertSame('Test', $view->render('file'));
-        $this->assertSame('Test', $view
-            ->withDefaultExtension('tpl')
-            ->render('file'));
-        $this->assertSame('Test', $view
-            ->withDefaultExtension('txt')
-            ->render('file'));
+        $this->assertSame('Test php', $view->render('file'));
         $this->assertSame(
-            'Test',
-            $view->withDefaultExtension('phpt')->withFallbackExtension('tpl')->render('file')
+            'Test tpl',
+            $view->withDefaultExtension('tpl')
+                ->render('file')
+        );
+        $this->assertSame(
+            'Test txt',
+            $view->withDefaultExtension('twig')
+                ->render('file.txt')
+        );
+        $this->assertSame(
+            'Test phpt',
+            $view->withDefaultExtension('twig')
+                ->withFallbackExtension('phpt')
+                ->render('file')
         );
     }