+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## 8.0.1 under development
 
 - Bug #224: Fix signature of `CachedContent::cache()` (@vjik)
+- Enh #232: Make fallback extension configurable (@rustamwin)
+- Chg #232: Deprecate `ViewInterface::withDefaultExtension()` and `ViewInterface::getDefaultExtension()` in favor of 
+  `ViewInterface::withFallbackExtension()` and `ViewInterface::getFallbackExtension()` (@rustamwin)
+- Bug #232: Fix render templates that contain dots in their name (@rustamwin)
 
 ## 8.0.0 February 16, 2023